update the order is valide price rather that using the product variant precentage with mutliple that with the  product price  i use that or the product price as the product price without mutliplication

+++ Infrastructure/Repositories/PaymentTypeRepository.cs
@@ -22,7 +22,7 @@ public class PaymentTypeRepository(AppDbContext context):IPaymentTypeRepository
         return await context.PaymentTypes.AsNoTracking().FirstOrDefaultAsync(x => x.Id == id);
     }
     
-    public async Task<List<PaymentType>> GetPaymentTypes(sbyte pageNum,sbyte pageSie)
+    public async Task<List<PaymentType>> GetPaymentTypes(int pageNum, int pageSie)
     {
         return await context.PaymentTypes.AsNoTracking()
             .Take(pageSie)
