+++ application/Interface/IPaymentTypeServices.cs
@@ -8,5 +8,5 @@ public interface IPaymentTypeServices
 {
     public Task<Result<PaymentTypeDto?>> Create(CreatePaymentTypeDto paymentTypeDto,Guid adminId);
     public Task<Result<PaymentTypeDto?>> Update(UpdatePaymentTypeDto paymentTypeDto, Guid adminId);
-    public Task<Result<List<PaymentTypeDto>?>> GetPaymentTypes(sbyte pageNum,sbyte pageSie=25);
+    public Task<Result<List<PaymentTypeDto>?>> GetPaymentTypes(int pageNum, int pageSie = 25);
 }
