+++ domain/entity/ProductVariant.cs
@@ -2,7 +2,7 @@ namespace api.domain.entity;
 
 public class ProductVariant:GeneralSharedInfoWithId
 {
-    public ushort Percentage { get; set; }
+    public int Percentage { get; set; }
     public Guid VariantId { get; set; }
     public string Name { get; set; }
     public Guid ProductId { get; set; }
